Reorganize http module before refactor
