--- packages/cli-kit/src/http/formdata.ts
@@ -1,7 +1,0 @@
-import FormData from 'form-data'
-
-function formData() {
-  return new FormData()
-}
-
-export default formData
