+++ packages/cli-kit/src/public/node/api/admin.ts
@@ -2,8 +2,7 @@ import {AdminSession} from '../session.js'
 import {content, token as outputToken} from '../../../output.js'
 import {Bug, Abort} from '../../../error.js'
 import {graphqlRequest, GraphQLVariables} from '../../../private/node/api/graphql.js'
-import {restRequestBody, restRequestHeaders, restRequestUrl} from '../../../private/node/api/rest.js'
-import fetch from '../../../http/fetch.js'
+import {fetch, restRequestBody, restRequestHeaders, restRequestUrl} from '../../../private/node/api/rest.js'
 import {ClientError, gql} from 'graphql-request'
 
 /**
